Rename UtfView and utfview to utf_view
This brings the project more into alignment with the Beman project's
preference for snake_case, which was added to the Beman project by the
following commit:

https://github.com/beman-project/beman/commit/7016bcc42096527d6b23f4828a77ae745247d6aa
