+++ .github/workflows/ci.yml
@@ -27,11 +27,11 @@ jobs:
       - name: Coverage
         run: |
           lcov --directory ./build --capture --output-file ./build/coverage_all.info
-          lcov --remove ./build/coverage_all.info -o ./build/coverage.info '/usr/include/*' "$PWD/src/UtfView/tests/*" "$PWD/deps/*"
+          lcov --remove ./build/coverage_all.info -o ./build/coverage.info '/usr/include/*' "$PWD/src/utf_view/tests/*" "$PWD/deps/*"
       - name: Coveralls
         uses: coverallsapp/github-action@master
         with:
-          path-to-lcov: ${{runner.workspace}}/UtfView/build/coverage.info
+          path-to-lcov: ${{runner.workspace}}/utf_view/build/coverage.info
           github-token: ${{ secrets.GITHUB_TOKEN }}
   clang-linux:
     runs-on: ubuntu-latest
