+++ README.md
@@ -7,9 +7,9 @@ Distributed under the Boost Software License, Version 1.0.
          https://www.boost.org/LICENSE_1_0.txt)
 -->
 
-# UtfView: C++26 UTF Transcoding Views
+# utf_view: C++26 UTF Transcoding Views
 
-[![CI](https://github.com/ednolan/UtfView/actions/workflows/ci.yml/badge.svg)](https://github.com/ednolan/utfview/actions) [![Coverage](https://coveralls.io/repos/github/ednolan/UtfView/badge.svg?branch=main)](https://coveralls.io/github/ednolan/UtfView?branch=main)
+[![CI](https://github.com/ednolan/utf_view/actions/workflows/ci.yml/badge.svg)](https://github.com/ednolan/utf_view/actions) [![Coverage](https://coveralls.io/repos/github/ednolan/utf_view/badge.svg?branch=main)](https://coveralls.io/github/ednolan/utf_view?branch=main)
 
 This repository implements the functionality proposed by P2728 for C++26, including:
 
@@ -23,7 +23,7 @@ Transcoding a UTF-8 string literal to a `std::u32string`:
 
 ```
 std::u32string hello_world =
-  u8"こんにちは世界" | utfview::to_utf32 | std::ranges::to<std::u32string>();
+  u8"こんにちは世界" | utf_view::to_utf32 | std::ranges::to<std::u32string>();
 ```
 
 Sanitizing potentially invalid Unicode C strings by replacing invalid code units with replacement characters:
@@ -31,7 +31,7 @@ Sanitizing potentially invalid Unicode C strings by replacing invalid code units
 ```
 template <typename CharT>
 std::basic_string<CharT> sanitize(CharT const* str) {
-  return utfview::null_term(str) | utfview::to_utf<CharT> | std::ranges::to<std::basic_string<CharT>>();
+  return utf_view::null_term(str) | utf_view::to_utf<CharT> | std::ranges::to<std::basic_string<CharT>>();
 }
 ```
 
@@ -39,7 +39,7 @@ Returning the final non-ASCII code point in a string, transcoding backwards lazi
 
 ```
 std::optional<char32_t> last_nonascii(std::ranges::view auto str) {
-  for (auto c : str | utfview::to_utf32 | std::views::reverse 
+  for (auto c : str | utf_view::to_utf32 | std::views::reverse 
                     | std::views::filter([](char32_t c) { return c > 0x7f; })
                     | std::views::take(1)) {
     return c;
@@ -75,9 +75,9 @@ std::basic_string<ToChar> transcode_or_throw(std::basic_string_view<FromChar> in
 
 ### Minimum Requirements
 
-UtfView requires being built in C++23 mode.
+utf_view requires being built in C++23 mode.
 
-UtfView has the following minimum compiler requirements:
+utf_view has the following minimum compiler requirements:
 
 - GCC 14
 - Clang 19
@@ -85,7 +85,7 @@ UtfView has the following minimum compiler requirements:
 
 ### Dependencies
 
-UtfView depends on [Boost.STLInterfaces](https://github.com/boostorg/stl_interfaces). It brings in this library and its dependencies via git submodules. If you add UtfView to your project your project already uses Boost, it will target your project's stl_interfaces instead of using its own submodule.
+utf_view depends on [Boost.STLInterfaces](https://github.com/boostorg/stl_interfaces). It brings in this library and its dependencies via git submodules. If you add utf_view to your project your project already uses Boost, it will target your project's stl_interfaces instead of using its own submodule.
 
 ### Instructions
 
@@ -99,10 +99,10 @@ This excerpt from the project's CI script provides an example of building the pr
 
 ### Paper
 
-UtfView is based on P2728.
+utf_view is based on P2728.
 
 - The latest official revision of P2728 can be found at https://wg21.link/p2728
-- The unofficial latest draft Markdown source for the paper can be found in this repository at [paper/P2828.md](https://github.com/ednolan/UtfView/blob/main/paper/P2728.md)
+- The unofficial latest draft Markdown source for the paper can be found in this repository at [paper/P2828.md](https://github.com/ednolan/utf_view/blob/main/paper/P2728.md)
 - The paper's committee status page can be found at https://github.com/cplusplus/papers/issues/1422
 
 ## Authors
@@ -111,4 +111,4 @@ This implementation of P2728 is a fork by Eddie Nolan of the implementation of P
 
 ## License
 
-UtfView is licensed under BSL 1.0.
+utf_view is licensed under BSL 1.0.
