--- include/UtfView/utfview.hpp
@@ -1,8 +1,0 @@
-#ifndef UTFVIEW_UTFVIEW_HPP
-#define UTFVIEW_UTFVIEW_HPP
-
-#include <UtfView/code_unit_view.hpp>
-#include <UtfView/null_term.hpp>
-#include <UtfView/to_utf_view.hpp>
-
-#endif UTFVIEW_UTFVIEW_HPP
