+++ paper/P2728.md
@@ -6128,7 +6128,7 @@ __Outcome:__ Strong consensus in favor
 # Implementation experience
 
 The most recent revision of this paper has a reference implementation called
-[UtfView](https://github.com/ednolan/utfview) available on GitHub, which is a
+[utf_view](https://github.com/ednolan/utf_view) available on GitHub, which is a
 fork of Jonathan Wakely's implementation of P2728R6 as an implementation
 detail for libstdc++.
