+++ src/utf_view/tests/CMakeLists.txt
@@ -6,7 +6,7 @@
 #          https://www.boost.org/LICENSE_1_0.txt)
 
 add_library(
-  utfview_test_lib
+  utf_view_test_lib
   STATIC
   code_unit_view.t.cpp
   detail/concepts.t.cpp
@@ -17,12 +17,12 @@ add_library(
   std_archetypes/iterator.t.cpp
   to_utf_view.t.cpp)
 
-target_link_libraries(utfview_test_lib utfview)
+target_link_libraries(utf_view_test_lib utf_view)
 
 add_executable(
-  utfview_test
+  utf_view_test
   main.t.cpp)
 
-target_link_libraries(utfview_test "$<LINK_LIBRARY:WHOLE_ARCHIVE,utfview_test_lib>")
+target_link_libraries(utf_view_test "$<LINK_LIBRARY:WHOLE_ARCHIVE,utf_view_test_lib>")
 
-add_test(NAME utfview_test COMMAND utfview_test)
+add_test(NAME utf_view_test COMMAND utf_view_test)
