+++ src/utf_view/tests/framework.hpp
@@ -12,7 +12,7 @@
 #include <map>
 #include <string>
 
-namespace utfview::tests {
+namespace utf_view::tests {
 
 namespace framework {
 
@@ -20,6 +20,6 @@ namespace framework {
 
 } // namespace framework
 
-} // namespace utfview::tests
+} // namespace utf_view::tests
 
 #endif // UTFVIEW_TESTS_FRAMEWORK_HPP
