+++ src/utf_view/tests/null_term.t.cpp
@@ -5,7 +5,7 @@
 //    (See accompanying file LICENSE.txt or copy at
 //          https://www.boost.org/LICENSE_1_0.txt)
 
-#include <UtfView/null_term.hpp>
+#include <utf_view/null_term.hpp>
 #include <concepts>
 #include <cstddef>
 #include <cstdint>
@@ -14,7 +14,7 @@
 #include <tests/std_archetypes/iterator.hpp>
 #include <utility>
 
-namespace utfview::tests {
+namespace utf_view::tests {
 
 static_assert(std::default_initializable<
               std::iter_value_t<std_archetypes::input_iterator_archetype>>);
@@ -52,4 +52,4 @@ constexpr bool null_sentinel_forward_iterator_test() {
 static_assert(null_sentinel_input_iterator_test());
 static_assert(null_sentinel_forward_iterator_test());
 
-} // namespace utfview::tests
+} // namespace utf_view::tests
